+++ src/proxy/http.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/httputil"
+	"github.com/skycoin/teller/src/util/httputil"
 	"github.com/skycoin/teller/src/logger"
 )
