+++ src/service/btcaddrs/store_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/boltdb/bolt"
 	"github.com/stretchr/testify/require"
 
-	"github.com/skycoin/teller/src/dbutil"
+	"github.com/skycoin/teller/src/util/dbutil"
 	"github.com/skycoin/teller/src/service/testutil"
 )
